# openapi/Iedora.Auth.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Iedora.Auth | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": [
          "Iedora.Auth"
        ],
        "summary": "Rotate the refresh-token cookie and issue a fresh access token.",
        "operationId": "Refresh",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/change-password": {
      "post": {
        "tags": [
          "Iedora.Auth"
        ],
        "summary": "Change the signed-in user's password; revokes the user's other sessions.",
        "operationId": "ChangePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChangePasswordRequest": {
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": [
              "null",
              "string"
            ]
          },
          "newPassword": {
            "maxLength": 128,
            "minLength": 8,
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TokenResponse": {
        "required": [
          "accessToken",
          "expiresAt",
          "userId",
          "tenantId",
          "mustChangePassword"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "tenantId": {
            "type": [
              "null",
              "string"
            ]
          },
          "mustChangePassword": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
    }
  },
...
}